Add files via upload
Hotspot analysis is a spatial analysis and mapping technique interested in the identification of clustering of spatial phenomena. These spatial phenomena are depicted as points in a map and refer to locations of events or objects. 

In this code, I have made use of the ESDA (Exploratory Spatial Data Analysis) concept and library to understand where the most fraudulent claims come from in the United States.